requirements.txt
streamlit==1.32.0
openai==0.28.0
python-dotenv==1.0.1